pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aisecops-interceptor"
version = "0.1.0"
description = "Runtime security interceptor for AI agents"
readme = "README.md"
requires-python = ">=3.11,<3.14"
authors = [{name = "Viplav Fauzdar"}]
dependencies = [
  "fastapi==0.116.1",
  "uvicorn==0.35.0",
  "pydantic==2.11.7",
  "PyYAML==6.0.2",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
